refactor(opensidian): extract workspace infra from fsState into workspace.ts
Split createFsState() into two modules so workspace infrastructure
(Yjs workspace, filesystem, extensions) lives in pure TypeScript
separate from Svelte reactive UI state. Consumers that only need
infra (Toolbar, ContentEditor) now import directly from workspace.ts.
